Update promotion usage count calculations
This changes the way that promotion usage counts work in a few ways.
For one, line item promotions are calculated as one usage per order.
This fixes a bunch of issues and is probably the way it should have
been handled by default anyway. It also fixes the usage count being
wrong for line items which have a quantity of more than one. In
addition to that, only completed orders now count towards the usage
limit.

This commit takes heavy inspiration, and basically backports the
changes, from an early solidus commit. See here:
https://github.com/solidusio/solidus/commit/58ee98992e2aa5f2000ecdebe11a91e77aa0b8f8